fantastic be fore we proceed lets fix this error
Console Error

React does not recognize the `indicatorClassName` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `indicatorclassname` instead. If you accidentally passed it from a parent component, remove it from the DOM element.

src/components/ui/progress.tsx (12:3) @ _c

  10 |   React.ComponentPropsWithoutRef<typeof ProgressPrimitive.Root>
  11 | >(({ className, value, ...props }, ref) => (
> 12 |   <ProgressPrimitive.Root
     |   ^
  13 |     ref={ref}
  14 |     className={cn(
  15 |       "relative h-4 w-full overflow-hidden rounded-full bg-secondary",